sync: Move hazard detection code
